// AAA-ANG-FRONTEND-ACCIONISTAS/src/app/model/titles.model.ts
export interface Titles {
    ACCIONISTA_UUID?: string;
    ACCIONES?: number;
    ACCIONES_TITULO?: number
    ACCIONES_RESTANTES?: number
    ACCIONES_TRASLADO?: number
    CEDULA?: string;
    CONSECUTIVO?: string;
    ESTADO?: string;
    ESTADO_UUID?: string;
    FECHA_ENTREGA?: string;
    FECHA_EXPEDICION?: string;
    FECHA_SUSCRIPCION?: string;
    IDENTIFICACION_ACCIONISTA?: number;
    NOMBRE_ACCIONISTA?: string;
    OBSERVACION?: number;
    TIPO_ACCIONISTA?: string;
    TITULO_UUID?: string;
    //rol y usuario del que hace peticion
    ROLADMIN?: number;
    EMAILADMIN?: string;
    CHECKED?: boolean;
}